Fix MinMaxSlider
As of Unity 6.0.23f, the stock component has a bug that sometimes causes error messages spam. It impacts FPS. This change introduces a copy of the component with the fixed method. As well as the TAPI preset for it.

Follow https://discussions.unity.com/t/unity-6-uitoolkit-layout-update-is-struggling-to-process-current-layout-with-minmaxslider-in-editor/1538464 to track when it's fixed in Unity and then rollback.